apollo_staking: create new crate and move the committee manager to it
This change includes moving around code to new crate and splitting it to various files, but there are no code changes.
