+++ commitlint.config.js
@@ -41,6 +41,7 @@ const AllowedScopes = ['apollo_batcher',
     'apollo_rpc_execution',
     'apollo_signature_manager',
     'apollo_signature_manager_types',
+    'apollo_staking',
     'apollo_starknet_client',
     'apollo_starknet_os_program',
     'apollo_state_reader',
