+++ crates/apollo_consensus_orchestrator/src/lib.rs
@@ -15,9 +15,6 @@ pub mod validate_proposal;
 #[allow(missing_docs)]
 pub mod cende;
 
-#[allow(missing_docs)]
-pub mod committee_manager;
-
 /// Fee market logic.
 pub mod fee_market;
