+++ crates/apollo_staking/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod committee_provider;
+pub mod contract_types;
+pub mod staking_manager;
+
+mod utils;
